percent

+++ Getdiskinfo.h
@@ -4,7 +4,7 @@
     extern "C" { /* 如果没有采用C++，顺序预编译 */
 #endif
 /* 采用C编译器编译的C语言代码段 */
-long get_disk_percent(const char* disk);
+int get_disk_percent(const char* disk);
 #ifdef __cplusplus /* 结束使用C编译器 */
     }
 #endif
